+++ src/main/java/io/sfrei/tracksearch/tracks/deserializer/youtube/YouTubeListTrackDeserializer.java
@@ -80,6 +80,8 @@ public class YouTubeListTrackDeserializer extends JsonDeserializer<YouTubeTrack.
 
         youTubeTrackBuilder.trackMetadata(new YouTubeTrackMetadata(channelName, channelUrl, streamAmount, thumbNailUrl));
 
+        // Info - cannot be obtained here
+
         return listYouTubeTrackBuilder;
     }