+++ src/main/java/io/sfrei/tracksearch/tracks/deserializer/youtube/YouTubeURLTrackDeserializer.java
@@ -75,6 +75,8 @@ public class YouTubeURLTrackDeserializer extends JsonDeserializer<YouTubeTrack.U
 
         youTubeTrackBuilder.trackMetadata(new YouTubeTrackMetadata(channelName, channelUrl, streamAmount, thumbNailUrl));
 
+        // Info - cannot be obtained here
+
         return listYouTubeTrackBuilder;
     }